プロフィール取得用のrepositotyを作成し、APIcall処理をserviceから分離
