+++ src/app/app.module.ts
@@ -6,6 +6,8 @@ import { ScullyLibModule } from '@scullyio/ng-lib';
 import { HeaderComponent } from './presentation/components/Header/Header.component';
 import { BlogsComponent } from './presentation/components/blogs/blogs.component';
 import { SharedModule } from './presentation/components/shared/Shared.module';
+import { ProfileRepository } from './domain/repositories/profilerepository/profile.repository';
+import { ProfileRepositoryImpl } from './infra/repositories/profilerepository/profile.repository';
 
 @NgModule({
   declarations: [
@@ -18,7 +20,8 @@ import { SharedModule } from './presentation/components/shared/Shared.module';
     ScullyLibModule,
     SharedModule
   ],
-  providers: [],
+  providers:
+    [{ provide: ProfileRepository, useClass: ProfileRepositoryImpl }],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
