+++ src/app/domain/repositories/profilerepository/profile.repository.ts
@@ -1,0 +1,13 @@
+import { Injectable } from '@angular/core';
+import { Observable } from 'rxjs';
+import { TakuminishProfile } from '../../model/resource-model/takuminishprofile.model';
+
+@Injectable({
+    providedIn: 'root'
+})
+export abstract class ProfileRepository {
+
+    constructor() { }
+
+    abstract fetchProfile(): Observable<TakuminishProfile>;
+}
