+++ src/app/presentation/components/top/top.component.ts
@@ -1,8 +1,8 @@
 import { Component, EventEmitter, OnDestroy, OnInit } from '@angular/core';
 import { isScullyGenerated } from '@scullyio/ng-lib';
 import { Observable, takeUntil } from 'rxjs';
+import { ProfileService } from 'src/app/domain/service/profileservice/profile.service';
 import { TakuminishProfile } from '../../../domain/model/resource-model/takuminishprofile.model';
-import { ProfileService } from '../../../domain/service/profile.service';
 
 @Component({
   selector: 'app-top',
