+++ src/app/presentation/components/top/top.module.ts
@@ -1,5 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
+import { ProfileRepository } from 'src/app/domain/repositories/profilerepository/profile.repository';
+import { ProfileRepositoryImpl } from 'src/app/infra/repositories/profilerepository/profile.repository';
 import { SharedModule } from '../shared/Shared.module';
 import { ExperienceComponent } from './Experience/Experience.component';
 import { LicenseComponent } from './License/License.component';
@@ -26,6 +28,7 @@ import { TopComponent } from './top.component';
         SkillComponent,
         LicenseComponent,
         ProductComponent
-    ]
+    ],
+
 })
 export class TopModule { }
